Done connecting posts & users, authorization and clearing post-user relations.

+++ controllers/feed.js
@@ -4,6 +4,7 @@ const path = require('path');
 const { validationResult } = require('express-validator');
 
 const Post = require('../models/post');
+const User = require('../models/user');
 
 exports.getPosts = (req, res, next) => {
     const currentPage = req.query.page || 1;
@@ -49,19 +50,27 @@ exports.createPost = (req, res, next) => {
     const imageUrl = req.file.path.replace("\\", "/");
     const title = req.body.title;
     const content = req.body.content;
+    let creator;
 
     const post = new Post({
         title: title,
         imageUrl: imageUrl,
         content: content,
-        creator: { name: 'Testing' }
+        creator: req.userId
     });
 
     post.save()
-        .then(result => {
+        .then(() => User.findById(req.userId))
+        .then(user => {
+            creator = user;
+            user.posts.push(post);
+            return user.save();
+        })
+        .then(() => {
             res.status(201).json({
                 message: 'Post created successfully!',
-                post: result
+                post: post,
+                creator: { _id: creator._id, name: creator.name }
             });
         })
         .catch(err => {
@@ -118,6 +127,12 @@ exports.updatePost = (req, res, next) => {
                 throw error;
             }
 
+            if (post.creator.toString() !== req.userId) {
+                const error = new Error('Not authorized.');
+                error.statusCode = 403;
+                throw error;
+            }
+
             if (imageUrl !== post.imageUrl) clearImage(post.imageUrl);
 
             post.title = title;
@@ -137,6 +152,7 @@ exports.updatePost = (req, res, next) => {
 
 exports.deletePost = (req, res, next) => {
     const postId = req.params.postId;
+
     Post.findById(postId)
         .then(post => {
             if (!post) {
@@ -144,14 +160,23 @@ exports.deletePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw error;
             }
+
+            if (post.creator.toString() !== req.userId) {
+                const error = new Error('Not authorized.');
+                error.statusCode = 403;
+                throw error;
+            }
+
             // Check logged in user
             clearImage(post.imageUrl);
             return Post.findByIdAndRemove(postId);
         })
-        .then(result => {
-            console.log(result);
-            res.status(200).json({ message: 'Deleted post.' });
+        .then(() => User.findById(req.userId))
+        .then(user => {
+            user.posts.pull(postId);
+            return user.save();
         })
+        .then(() => res.status(200).json({ message: 'Deleted post.' }))
         .catch(err => {
             if (!err.statusCode) err.statusCode = 500;
             next(err);
